fix(esp/zephyr): logging

+++ src/platform/esp/logging.c
@@ -26,7 +26,7 @@ TINY_TOOL_WEAK void tinyPlatLog(tinyLogLevel log_level, const char *logRegion, c
     case TINY_LOG_LEVEL_CRIT:
         if (LOG_LOCAL_LEVEL >= ESP_LOG_ERROR)
         {
-            esp_log_write(ESP_LOG_ERROR, logRegion, LOG_COLOR_E "E(%u) %s:", esp_log_timestamp(), logRegion);
+            esp_log_write(ESP_LOG_ERROR, logRegion, LOG_COLOR_E "E(%lu) %s:", (long int)esp_log_timestamp(), logRegion);
             esp_log_writev(ESP_LOG_ERROR, logRegion, format, args);
             esp_log_write(ESP_LOG_ERROR, logRegion, LOG_RESET_COLOR "\n");
         }
@@ -34,7 +34,7 @@ TINY_TOOL_WEAK void tinyPlatLog(tinyLogLevel log_level, const char *logRegion, c
     case TINY_LOG_LEVEL_WARN:
         if (LOG_LOCAL_LEVEL >= ESP_LOG_WARN)
         {
-            esp_log_write(ESP_LOG_WARN, logRegion, LOG_COLOR_W "W(%u) %s:", esp_log_timestamp(), logRegion);
+            esp_log_write(ESP_LOG_WARN, logRegion, LOG_COLOR_W "W(%lu) %s:", (long int)esp_log_timestamp(), logRegion);
             esp_log_writev(ESP_LOG_WARN, logRegion, format, args);
             esp_log_write(ESP_LOG_WARN, logRegion, LOG_RESET_COLOR "\n");
         }
@@ -43,7 +43,7 @@ TINY_TOOL_WEAK void tinyPlatLog(tinyLogLevel log_level, const char *logRegion, c
     case TINY_LOG_LEVEL_INFO:
         if (LOG_LOCAL_LEVEL >= ESP_LOG_INFO)
         {
-            esp_log_write(ESP_LOG_INFO, logRegion, LOG_COLOR_I "I(%u) %s:", esp_log_timestamp(), logRegion);
+            esp_log_write(ESP_LOG_INFO, logRegion, LOG_COLOR_I "I(%lu) %s:", (long int)esp_log_timestamp(), logRegion);
             esp_log_writev(ESP_LOG_INFO, logRegion, format, args);
             esp_log_write(ESP_LOG_INFO, logRegion, LOG_RESET_COLOR "\n");
         }
@@ -51,7 +51,7 @@ TINY_TOOL_WEAK void tinyPlatLog(tinyLogLevel log_level, const char *logRegion, c
     default:
         if (LOG_LOCAL_LEVEL >= ESP_LOG_DEBUG)
         {
-            esp_log_write(ESP_LOG_DEBUG, logRegion, LOG_COLOR_D "D(%u) %s:", esp_log_timestamp(), logRegion);
+            esp_log_write(ESP_LOG_DEBUG, logRegion, LOG_COLOR_D "D(%lu) %s:", (long int)esp_log_timestamp(), logRegion);
             esp_log_writev(ESP_LOG_DEBUG, logRegion, format, args);
             esp_log_write(ESP_LOG_DEBUG, logRegion, LOG_RESET_COLOR "\n");
         }
